Move activities to its own folder
